Revert "1 and 2 SOLID principles applied to all Components folder"
This reverts commit 1e1438482ee0773875968b4a08f5adcff9cc4115.
